requirements.txt
torch==2.2.2
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
albumentations==2.0.0
loguru==0.7.3
scikit-learn==1.5.1
torchvision==0.17.2
pillow==10.4.0
tqdm==4.66.5
matplotlib==3.9.2
hydra-core==1.3.2
omegaconf==2.3.0